:art: Moved STL-related scripts to a separate folder.
:bug: :lipstick: Fixed progress bar.

:lipstick:

Fixed saving as prefab

:sparkles: Implemented importing and exporting name of the mesh

:construction: Fixing shading
