--- unity/Holo/Assets/Hypertrophy_test.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e812f79093f6d2e4b844086f0b789f7f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
