--- unity/Holo/Assets/Hypertrophymesh_test.mesh.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a5e77053ecc76314788f21133dbf5e7f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
