+++ unity/Holo/Assets/Scripts/STLConversion.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71fc20c69a7845544bd0d82ec41dd2b9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
