+++ unity/Holo/Assets/Scripts/STLConversion/STLConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6fa252e21c4931443adf98bd7c9ee563
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
