+++ unity/Holo/Assets/Scripts/STLConversion/STLFIleImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d5a7a165b1d234459f7c9f1807b4339
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
