+++ unity/Holo/Assets/Scripts/STLConversion/STLImportUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1acbd40d9afd7c42b8a42b868709f1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
