+++ unity/Holo/Assets/Scripts/STLConversion/STLSeriesImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e9edfd06d298634f86f6346ba269557
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
