--- unity/Holo/Assets/Scripts/STLConverter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40d7539493f19b2429c22a06b04b9245
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
