--- unity/Holo/Assets/Scripts/STLFIleImporter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae967410c71eb3e45acbefc18e148dbb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
